feat: making a page instead of a modal
