--- app/components/events-page/event-modal-with-onClose.tsx
@@ -1,12 +1,0 @@
-// "use client"
-
-// import EventModal from "@/app/components/events-page/event-modal";
-// import { Event } from "@/app/lib/types";
-// import { useRouter } from "next/navigation";
-
-// export default function EventModalWithOnClose({event}: {event: Event}) {
-//   const router = useRouter()
-//   return (
-//     <EventModal event={event} onClose={() => router.push("/events")}/>
-//   )
-// }
