+++ app/events/[id]/page.tsx
@@ -1,10 +1,8 @@
-import EventPage from "../page"
 import Modal from "./modal"
 
 export default function Page() {
   return (
     <>
-      <EventPage/>
       <Modal/>
     </>
   )
